% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_xlex.R
\name{demo_xlex}
\alias{demo_xlex}
\title{Run a shiny app to demonstrate xlex()}
\usage{
demo_xlex()
}
\description{
Requires the \code{shiny}, \code{igraph} and \code{ggraph} packages, which you
can install with \code{install.packages(c("shiny", "igraph", "ggraph"))}.  Runs a
shiny app to demonstrate tokenizing an Excel formula with \code{xlex()} and
plotting the parse tree with \code{plot_xlex()}
}
\examples{
if (interactive()
    && require(shiny, quietly = TRUE)
    && require(igraph, quietly = TRUE)
    && require(ggraph, quietly = TRUE)) {
  demo_xlex()
}
}
\seealso{
\code{\link[tidyxl:xlex]{tidyxl::xlex()}}, \code{\link[tidyxl:plot_xlex]{tidyxl::plot_xlex()}}
}
