% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend.R
\name{extend}
\alias{extend}
\title{Extend a bag of cells in one direction, optionally up to a boundary
condition.}
\usage{
extend(bag, cells, direction, boundary = FALSE, include = FALSE)
}
\arguments{
\item{bag}{Data frame. The original selection, including at least the columns
'row' and 'column', which are numeric/integer vectors.}

\item{cells}{Data frame. All the cells in the sheet, within which to extend
the bag. Including at least the columns 'row' and 'column', as well as any
columns referred to by the bounadry formula.}

\item{direction}{Character vector length 1. The direction in which to extend,
among the compass directions "N", "E", "S", "W", where "N" is north (up).}

\item{boundary}{Formula to express a boundary condition. Defaults to
`FALSE`, which means the extension will go to the boundary of the sheet. `~
col <= 50` would up to the 50th column, but see 'details' for how this might
behave unpredictably depending on 'include'.}

\item{include}{Logical vector length 1. Whether to include in the extension
the first cell at which the boundary condition is met.  Can be unpredictable
when `TRUE` if `boundary` is something like `~ col <= 50`, because if there
is no cell in the 50th column, then the first cell beyond the 50th column
will be included.}
}
\description{
A bag of data cells is a data frame with at least the
columns 'row' and 'col', as well as any others that carry information about
the cells, e.g. their values.  More cells may be added to the bag by
extending the bag in a given direction.  Unless a boundary condition is
specified, all cells up to the edge of the sheet will be added.
}
\details{
A bag may have ragged rows or ragged cols. Gaps will not be filled
in, and the edge that is extended will be justified up to the first
row/column in which the boundary condition is met.
}

