% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{formats}
\alias{formats}
\title{Import xlsx (Excel) formatting into a list structure for lookups.}
\usage{
formats(path)
}
\description{
\code{formats} imports formatting information from xlsx (Excel) spreadsheets
into a list structure.  The top level of the list is divided into two parts:
\code{$style}, which contains formats applied in a general way by cell
'styles', and \code{$local}, which contains formats applied to individual
cells.  This structure reflects how a cell may be formatted according to a
general 'style', yet also differ from the 'style' by some local modification, 
e.g. by making it bold.
}
\details{
\code{\link{contents}} Returns, among other things, keys for each cell into
the \code{$style} and \code{$local} substructures of \code{formats}' output.
These keys can be used to retrieve the style of each cell by subsetting the
relevant part of \code{formats}' output.  See how this is done in the 
examples.

Colours may be recorded in any of three ways: a hexadecimal RGB string with
or without alpha, an 'indexed' colour, and an index into a 'theme'.
\code{formats} dereferences 'indexed' and 'theme' colours to their hexadecimal
RGB string representation, and standardises all RGB strings to have an alpha
channel in the first two characters.  The 'index' and the 'theme' name are
still provided.  To filter by an RGB string, you might to look up the RGB 
values in a spreadsheet program (e.g. Excel, LibreOffice, Gnumeric), and use 
the \code{\link{rgb}} function to convert these to a hexadecimal string.

\code{formats} is currently quite slow for spreadsheets with complex
formatting.  It is planned to implement \code{formats} in C++ to speed it up.
}
\examples{
examples <- system.file("extdata/examples.xlsx", package = "tidyxl")
examples_formats <- formats(examples)
str(examples_formats)

# The formats of particular cells can be retrieved like this:
sheet1 <- contents(examples, "Sheet1")[[1]]
examples_formats$style$font$bold[sheet1$style_format_id]
examples_formats$local$font$bold[sheet1$local_format_id]

# To filter for cells of a particular format, first filter the formats to get
# the relevant indices, and then filter the cells by those indices.
bold_indices <- which(examples_formats$local$font$bold)
sheet1[sheet1$local_format_id \%in\% bold_indices, ]
}

