% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_date_format.R
\name{is_date_format}
\alias{is_date_format}
\title{Test that Excel number formats are date formats}
\usage{
is_date_format(x)
}
\arguments{
\item{x}{character vector of number format strings}
}
\description{
Test that Excel number formats are date formats
}
\examples{
is_date_format(c("yyyy-mm-dd", "0.0\%", "h:m:s", "£#,##0;[Red]-£#,##0"))
}
