% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_xlex.R
\name{plot_xlex}
\alias{plot_xlex}
\alias{xlex_edges}
\alias{xlex_vertices}
\alias{xlex_igraph}
\title{Plot the parse tree of a formula}
\usage{
plot_xlex(x)

xlex_edges(x)

xlex_vertices(x)

xlex_igraph(x)
}
\arguments{
\item{x}{parse tree from \code{xlex()}}
}
\description{
Requires the \code{igraph} and \code{ggraph} packages, which you
can install with \code{install.packages(c("igraph", "ggraph"))}.  Draws a
simple tree plot of the parse tree, showing parent-child relationships,
tokens and types.

The utility functions \code{xlex_edges()}, \code{xlex_vertices()}, and \code{xlex_igraph()}
compute each stage in plotting the final graph.
}
\examples{
x <- xlex("MIN(3,MAX(2,A1))")

xlex_edges(x)
xlex_vertices(x)

if (interactive() && require(igraph, quietly = TRUE)) {
  xlex_igraph(x)
}

if (interactive()
    && require(igraph, quietly = TRUE)
    && require(ggraph, quietly = TRUE)) {
  plot_xlex(x)
}
}
\seealso{
\code{\link[tidyxl:xlex]{tidyxl::xlex()}}, \code{\link[tidyxl:demo_xlex]{tidyxl::demo_xlex()}}
}
