% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyxl-package.R
\docType{package}
\name{tidyxl}
\alias{tidyxl}
\alias{tidyxl-package}
\title{tidyxl: Import xlsx (Excel) spreadsheet data and formatting into tidy structures.}
\description{
Tidyxl imports data from spreadsheets without coercing it into a rectangle,
and retains information encoded in cell formatting (e.g. font/fill/border).
This data structure is compatible with the 'unpivotr' package for
recognising information expressed by relative cell positions and cell
formatting, and re-expressing it in a tidy way.
}
\section{Spreadsheet functions}{

\itemize{
\item \code{\link[tidyxl:tidy_xlsx]{tidyxl::tidy_xlsx()}} Import cells and formatting from an xlsx file.
\item \code{\link[tidyxl:xlsx_sheet_names]{tidyxl::xlsx_sheet_names()}} List the names of sheets in an xlsx file.
\item \code{\link[tidyxl:xlsx_names]{tidyxl::xlsx_names()}} Import names and definitions of named ranges (aka named
\item \code{\link[tidyxl:is_range]{tidyxl::is_range()}} Test whether a 'name' from \code{\link[tidyxl:xlsx_names]{tidyxl::xlsx_names()}}
refers to a range or not.
formulas, defined names) from an xlsx file.
\item \code{\link[tidyxl:xlsx_validation]{tidyxl::xlsx_validation()}} Import cell input validation rules (e.g.
'must be from this drop-down list') from an xlsx file.
}
}

\section{Formula functions}{

\itemize{
\item \code{\link[tidyxl:xlex]{tidyxl::xlex()}} Tokenise (lex) an Excel formula.
\item \code{\link[tidyxl:plot_xlex]{tidyxl::plot_xlex()}} Draw a simple tree plot of a parse tree from
\code{\link[tidyxl:xlex]{tidyxl::xlex()}}
\item \code{\link[tidyxl:xlex_edges]{tidyxl::xlex_edges()}} Utility function used by \code{\link[tidyxl:plot_xlex]{tidyxl::plot_xlex()}}
\item \code{\link[tidyxl:xlex_vertices]{tidyxl::xlex_vertices()}} Utility function used by \code{\link[tidyxl:plot_xlex]{tidyxl::plot_xlex()}}
\item \code{\link[tidyxl:xlex_igraph]{tidyxl::xlex_igraph()}} Utility function used by \code{\link[tidyxl:plot_xlex]{tidyxl::plot_xlex()}}
}
}

