% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xlsx_color_standard}
\alias{xlsx_color_standard}
\alias{xlsx_colour_standard}
\title{Names and RGB values of Excel standard colours}
\format{A data frame with 10 rows and 2 variables:
\itemize{
\item \code{name} Name of the colour
\item \code{rgb}  RGB value of the colour
}}
\usage{
xlsx_color_standard

xlsx_colour_standard
}
\description{
A dataset containing the names and RGB colour values of Excel's standard
palette.
}
\keyword{datasets}
