% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_header.R
\name{join_header}
\alias{ABOVE}
\alias{LEFT}
\alias{N}
\alias{NNW}
\alias{W}
\alias{WNW}
\alias{join_header}
\title{Join a bag of data cells to some header, by proximity in a given direction, e.g. NNW is up and up-left.}
\usage{
join_header(bag, header, direction, colname = "newheader")

N(bag, header, colname = "newheader")

W(bag, header, colname = "newheader")

NNW(bag, header, colname = "newheader")

WNW(bag, header, colname = "newheader")

ABOVE(bag, header, colname = "newheader")

LEFT(bag, header, colname = "newheader")
}
\arguments{
\item{bag}{Data frame. A bag of data cells including at least the columns
'row' and 'column', which are numeric/integer vectors.}

\item{header}{Data frame. A bag of header cells, which currently must have
the columns 'row', 'col' and 'header', where 'header' is a column whose value
describes the data.  For example, for data cells recording a number of
animals, and a header describing the type of animal, then the 'header' column
might contain the values `c("dog", "cat", "mouse")`.}

\item{direction}{Character vector length 1. A compass direction to search for the nearest header.  See
'details'.}

\item{colname}{Character vector length one. Column name to give the header
values once they are joined to the 'bag'.  Continuing the example, this might
be "animal_type".}
}
\description{
A bag of data cells is a data frame with at least the
columns 'row' and 'col', as well as any others that carry information about
the cells, e.g. their values.  Cells in a table are associated with header
cells by proximity.  Having collected header cells and data cells into
separate data frames, 'join_header' and the related functions 'NNW', 'ABOVE',
etc., join the values in the header cells to the data cells, choose the
nearest header to each cell, in a given direction.
}
\details{
Headers are associated with data by proximity in a given direction.
The directions are mapped to the points of the compass, where 'N' is north
(up), 'E' is east (right), and so on.  `join_header` finds the nearest header
to a given data cell in a given direction, and joins its value to the data
cell.  The most common directions to search are 'NNW' (for left-aligned
headers at the top of the table) and 'WNW' for top-aligned headers at the
side of the table.  There can be a tie in the directions 'ABOVE', 'BELOW',
'LEFT' and 'RIGHT' (for headers that are not aligned to the edge of the data
cells that they refer to), and currently ties will cause the affected cells
not to be returned at all.  The full list of available directions is 'N',
'E', 'S', 'W', 'NE', 'NE', 'SE', 'SW', 'NNW', 'NNE', 'ENE', 'ESE', 'SSE',
'SSW', 'WSW', 'WNW'.  For convenience, these directions are provided as their
own functions, wrapping the concept of 'join_header'.
}
\section{Functions}{
\itemize{
\item \code{N}: Join nearest header in the 'N' direction.

\item \code{W}: Join nearest header in the 'W' direction.

\item \code{NNW}: Join nearest header in the 'NNW' direction.

\item \code{WNW}: Join nearest header in the 'WNW' direction.

\item \code{ABOVE}: Join nearest header in the 'ABOVE' direction.

\item \code{LEFT}: Join nearest header in the 'LEFT' direction.
}}

