% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyxl-package.r, R/tidyxl.R
\docType{package}
\name{tidyxl}
\alias{tidyxl}
\alias{tidyxl-package}
\title{tidyxl.}
\usage{
tidyxl(path, sheet = NA)
}
\arguments{
\item{path}{Path to the xlsx file}

\item{sheet}{Sheet to read. Either a string (the name of a sheet), or
an integer (the position of the sheet). Defaults to the first sheet.}
}
\description{
tidyxl.

Read xlsx data and formatting
}
\section{Functions}{
\itemize{
\item \code{tidyxl}: Read xlsx data and formatting semantically (not as a table)
}}
\examples{
datasets <- system.file("extdata/datasets.xlsx", package = "readxl")

# All sheets
tidyxl(datasets)

# Specific sheet either by position or by name
tidyxl(datasets, 2)
tidyxl(datasets, "mtcars")
}

